Moved cache to its own module
