+++ src/cache/cache.rs
@@ -1,3 +1,4 @@
+use crate::cache::inner_cache::CachedWithCode;
 use crate::config::redis_scan_count;
 use crate::utils::errors::{ApiError, ApiResult};
 use mockall::automock;
@@ -165,44 +166,6 @@ pub trait CacheExt: Cache {
         error_cache_duration: usize,
     ) -> ApiResult<String> {
         self.request_cached_advanced(client, url, cache_duration, error_cache_duration, false, 0)
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub(super) struct CachedWithCode {
-    pub(super) code: u16,
-    pub(super) data: String,
-}
-
-impl CachedWithCode {
-    const SEPARATOR: &'static str = ";";
-
-    pub(super) fn split(cached: &str) -> Self {
-        let cached_with_code: Vec<&str> = cached.splitn(2, CachedWithCode::SEPARATOR).collect();
-        CachedWithCode {
-            code: cached_with_code
-                .get(0)
-                .expect("Must have a status code")
-                .parse()
-                .expect("Not a valid Http code"),
-            data: cached_with_code.get(1).expect("Must have data").to_string(),
-        }
-    }
-
-    pub(super) fn join(code: u16, data: &str) -> String {
-        format!("{}{}{}", code, CachedWithCode::SEPARATOR, data)
-    }
-
-    pub(super) fn is_error(&self) -> bool {
-        200 > self.code || self.code >= 400
-    }
-
-    pub(super) fn to_result(&self) -> Result<String, ApiError> {
-        if self.is_error() {
-            Err(ApiError::from_backend_error(self.code, &self.data))
-        } else {
-            Ok(String::from(&self.data))
-        }
     }
 }
