+++ src/cache/mod.rs
@@ -1,0 +1,5 @@
+pub mod cache;
+mod inner_cache;
+
+#[cfg(test)]
+mod tests;
