+++ src/cache/tests/mod.rs
@@ -1,0 +1,1 @@
+mod cache;
