+++ src/routes/about.rs
@@ -1,7 +1,7 @@
+use crate::cache::cache::{Cache, CacheExt};
 use crate::config::base_transaction_service_url;
 use crate::config::{about_cache_duration, webhook_token};
 use crate::services::about;
-use crate::utils::cache::{Cache, CacheExt};
 use crate::utils::context::Context;
 use crate::utils::errors::ApiResult;
 use rocket::response::content;
