+++ src/routes/health.rs
@@ -1,5 +1,5 @@
+use crate::cache::cache::CacheExt;
 use crate::config::request_cache_duration;
-use crate::utils::cache::CacheExt;
 use crate::utils::context::Context;
 use crate::utils::errors::ApiResult;
 use rocket::response::content;
