+++ src/services/transactions_proposal.rs
@@ -1,6 +1,6 @@
+use crate::cache::cache::Cache;
 use crate::config::base_transaction_service_url;
 use crate::models::service::transactions::requests::MultisigTransactionRequest;
-use crate::utils::cache::Cache;
 use crate::utils::context::Context;
 use crate::utils::errors::{ApiError, ApiResult};
 use std::collections::HashMap;
