+++ src/utils/tests/mod.rs
@@ -1,4 +1,3 @@
-mod cache;
 mod data_decoded_utils;
 mod errors;
 mod json;
