+++ src/utils/transactions.rs
@@ -1,8 +1,8 @@
+use crate::cache::cache::CacheExt;
 use crate::config::{
     base_transaction_service_url, request_cache_duration, request_error_cache_timeout,
 };
 use crate::models::backend::transactions::MultisigTransaction;
-use crate::utils::cache::CacheExt;
 use crate::utils::context::Context;
 use ethabi::Uint;
 use ethcontract_common::hash::keccak256;
